Fix: [GitHub Actions] make sure apt-database is in sync
The OSes GitHub Actions run on, don't 'apt update' before becoming
available, and those OSes are not updated on a very regular base.
In result, hoping that a 'apt install' works is failing over time,
till the OS is updated again, and it slowly drifts into despair
again.

All this solved by an 'apt update' of course.