Fix: [Makefile] make sure installed filenames are as OpenTTD expects
It looks for readme.txt, license.txt and changelog.txt by name. So no matter
how we call it in the repository, make sure it matches with the expectation
in the installed directory.

This fix was already applied for the tarball in PR #24, but that missed
the filenames used when installing into the filesystem directly. See
also issue #22.