Upgrade #94: [Makefile] GIMP3 compatibility
 - Specify the batch script interpreter to use; GIMP3 provides
   multiple.

 - Use `gimp-console` instead of `gimp` to avoid a non-interactive
   mode issue[1].

This doesn't solve all compatibility problems with GIMP3 yet; we
also need to resolve the removal of a `gimp-image-get-active-layer`
API function in GIMP3.

The changes* here are intended to be backwards-compatible to GIMP2
and have been tested GIMP2.10.34 on Debian stable (bookworm), and
also with GIMP3.0.0 RC2 to build the OpenGFX 7.1 sources.

- [1] https://gitlab.gnome.org/GNOME/gimp/-/issues/12042

*NB: the `which` command has been superseded by `command -v` since
 v7.1 in VCS, so a small patch adjustment is required if backporting.