Fix #94: compatibility with gimp v2 again
This adds a second version of the gimpscript file containing the script
for gimp v2 (i.e. the version as it was before the previous commit). The
makefile now checks the gimp version and selects which of the two script
files to use based on that.

It would be nicer if the script itself checked the gimp version and
adapted accordingly, but it seems gimp v3 uses a newer scheme version
and newer syntactical elements (i.e. "#:foo" symbols) that the version
in gimp v2 does not understand, making this hard or impossible (at least
for an inexperienced scheme programmer).

This commit is intentionally separate from the previous commit (which
adapts the script for v3 compatibility), so the diff for the previous
commit is easier to read.

This is the last part to fix #94.